requirements.txt
# 文件解析
python-docx>=1.1.0
python-pptx>=0.6.23

# AI API
google-generativeai>=0.3.0

# Web 框架
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-Limiter>=3.5.0

# 資料驗證
pydantic>=2.5.0

# 任務存儲（可選，生產環境建議安裝）
redis>=5.0.0

# 工具庫
tqdm>=4.66.0
python-dotenv>=1.0.0
